+++ Tests/Unit/DataStructures/AbstractReadOnlyObjectWithAccessorsTest.php
@@ -287,7 +287,7 @@ class AbstractReadOnlyObjectWithAccessorsTest extends UnitTestCase
     }
 
     /**
-     * @return array<string, array>
+     * @return array<string, array{0: bool|int|string, 1: bool}>
      */
     public function booleanDataProvider(): array
     {
@@ -339,7 +339,7 @@ class AbstractReadOnlyObjectWithAccessorsTest extends UnitTestCase
     }
 
     /**
-     * @return array<string, array>
+     * @return array<string, array{0: float|string|int|bool, 1: float}>
      */
     public function floatDataProvider(): array
     {